chore: :wrench: add Sass architecture (#2)
* feat: :sparkles: add custom Sass variables and overrides for Bootstrap

This file serves as the single source of truth for theme-level design values, allowing for easy customization and theming of the application.

* feat: :sparkles: enhance styles.scss with detailed documentation

Added comprehensive comments to styles.scss to clarify the purpose, build output, and structure of the Sass entry point for better maintainability and understanding.

* feat: :sparkles: add app.scss for custom styles and extensions

Create app.scss as a custom extension layer for the design system, defining additional Sass components, utilities, and brand-specific styles. Update styles.scss to import app.scss for project-specific styles.