docs: :books: documentation enhancements (#6)
* docs: :books: add license section to README

Included a license section to clarify the project's licensing under the MIT license for private and commercial use.

* docs: :books: update introduction section in README

Expanded the introduction to provide a clearer foundation for building theme-driven CSS using Bootstrap and Sass.

* docs: :books: update README to include detailed features section

Added a comprehensive features section to the README, outlining the tools and technologies included in the template for building consistent CSS themes.

* docs: :books: add "Getting Started" section to README

This update introduces a new section in the README to guide users on how to quickly get started with the template.

* docs: :books: update README to include links in the template description

Enhanced the template description in the README to include hyperlinks for Bootstrap and Sass for better accessibility and reference.